### 2019-08-21 #1672 @kyleniemeyer
@yangkcatiu I think this PR should fix the paper compilation issue: https://github.com/IUNetSci/BotSlayer-CE/pull/11
