### 2019-08-29 #1672 @labarba
👋  @yangkcatiu
The README says:
> The "CE" version of BotSlayer is not meant for research
... which would imply this software is not in-scope for a JOSS publication. JOSS explicitly publishes research software.
