### 2019-08-29 #1672 @filmenczer
We apologize for the misleading warning. We eliminated it and updated the paper to clarify that the software is indeed ready for research, but depending on the specific application domain, users may wish to plug in more appropriate (simpler or more sophisticated) bot detection algorithms.

We can also suggest a few potential reviewers: @mkearney, @jboynyc, @soodoku, @rmhogervorst, @Ohmaddi.
