### 2019-08-30 #1672 @labarba
👋 @arokem — Could you handle this submission as editor?
"BotSlayer: real-time detection of bot amplification on Twitter"
Repository: https://github.com/IUNetSci/BotSlayer-CE
