### 2019-09-01 #1672 @arokem
👋 @mkearney : would you be willing to review this article for the [Journal of Open Source Software](https://joss.theoj.org/)?
