### 2019-09-01 #1672 @arokem
@nkratzke : thanks! Once I have one more reviewer, I will start the review as a separate issue here. You will then have a list of items to check off. For more information, please take a look at the [reviewer guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
