### 2019-09-03 #1672 @arokem
👋 @jboynyc : would you be willing to review this article for the [Journal of Open Source Software](https://joss.theoj.org/)?
