### 2019-09-03 #1706 @whedon
[REVIEW]: BotSlayer: real-time detection of bot amplification on Twitter
**Submitting author:** @yangkcatiu (<a href="http://orcid.org/0000-0003-4627-9273">Kai-Cheng Yang</a>)
**Repository:** <a href="https://github.com/IUNetSci/BotSlayer-CE" target ="_blank">https://github.com/IUNetSci/BotSlayer-CE</a>
**Version:** v1.0
