### 2019-09-04 IUNetSci/BotSlayer-CE#13 @nkratzke
JOSS Review: General remarks
I perform a review for JOSS and @arokem. Generally speaking, the software was easy to install using Docker and worked the first time without hassle on my Mac (iMac, Retina 5K, 27-inch, 2017. 3,8 GHz Intel Core i5, 32 GB 2400 MHz DDR4, macOS Version 10.14.6 (18G87)). It was easy to run an initial example case using the searchterm "Trump".

However, according to the JOSS checklist, there are open points from my point of view. These points are mainly documentation and automated testing points. I will address the following points in separate tickets.

- __List of dependencies:__ Ideally dependencies should be handled with an automated package management solution.
- __Example usage:__ Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- __Functionality documentation:__ Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
- __Automated tests:__ Are there automated tests or manual steps described so that the functionality of the software can be verified?
- __Community guidelines:__ Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
