### 2019-09-04 IUNetSci/BotSlayer-CE#14 @nkratzke
JOSS Review: List of dependencies
Ideally these should be handled with an automated package management solution, this is mainly done in your Docker setting using Python3/NodeJS and the pip/npm package management system.

However, in case of the Postgres database you are relying on the latest Docker image of Postgres SQL according to your Dockerfile.

`FROM library/postgres`

You should add the latest working version here to avoid image build breakdowns of prospective Postgres updates.

Additionally, you could provide a docker image on the Dockerhub for even better user convenience (and simplified installation, the image build took a while).
