### 2019-09-04 IUNetSci/BotSlayer-CE#15 @nkratzke
JOSS Review: Example usage
According to JOSS guidelines authors should include examples of how to use the software (ideally to solve real-world analysis problems).

The software is not complex to use, however, I missed a small tutorial with a real-world use case. You could add this easily.

What is more, I needed two approaches to figuring out how to insert search terms into the bot detecting engine. A tutorial would be helpful here.
