### 2019-09-04 IUNetSci/BotSlayer-CE#16 @nkratzke
JOSS Review: API Documentation
According to JOSS guidelines the core functionality of the software should be documented to a satisfactory level (e.g., API method documentation)?

I observed the following two points:

- I found no noteworthy API documentation (only some very basic inline comments).
- You architecture figure could be a bit more explicit aligned to your general code organization (backend, middleware, frontend).
