### 2019-09-04 IUNetSci/BotSlayer-CE#17 @nkratzke
JOSS Review: Automated tests
According to JOSS guidelines, there should be automated tests or manual steps described so that the functionality of the software can be verified.

I found neither the description of manual test steps nor automated tests in your repository. You may think about adding automated tests and report on code coverage to have an objective quality criterion?!
