### 2019-09-04 IUNetSci/BotSlayer-CE#18 @nkratzke
JOSS Review: Community guidelines
According to JOSS guidelines, community guidelines should be provided. These should cover guidelines for third parties wishing to

1. contribute to the software,
2. report issues or problems with the software,
3. seek support.
