### 2019-09-04 #1706 @nkratzke
@arokem I have done my review. Generally speaking, the software was easy to install using Docker and worked the first time without hassle. It was easy to run an initial example case using the searchterm "Trump".

However, according to the JOSS checklist, there are open points from my point of view. These points are mainly documentation and automated testing points. I have addressed these points in separate tickets

- https://github.com/IUNetSci/BotSlayer-CE/issues/13
- https://github.com/IUNetSci/BotSlayer-CE/issues/14
- https://github.com/IUNetSci/BotSlayer-CE/issues/15
- https://github.com/IUNetSci/BotSlayer-CE/issues/16
- https://github.com/IUNetSci/BotSlayer-CE/issues/17
- https://github.com/IUNetSci/BotSlayer-CE/issues/18

in the respective [repository](https://github.com/IUNetSci/BotSlayer-CE). The tickets cover mainly documentation points of the repository.

- Example usage
- Functionality documentation (missing API documentation)
- Automated tests (no automated tests found)
- Community guidelines (no guidelines found)

However, the JOSS paper is fine from my point of view.
