### 2019-09-06 IUNetSci/BotSlayer-CE#16 @marcmccarty
Commit c782feccc7ee7b9e6acbfbc77d693055f3c40d4d contains more in-depth comments on how the API/middleware functions.

Please let me know if there should be more info in there and/or somewhere else, but I think it's probably sufficient between the new comments and the README.
