### 2019-09-06 IUNetSci/BotSlayer-CE#16 @nkratzke
Content is fine. However, the API should be documented in a standalone way. A best practice would be to generate the API documentation from the API method comments.
