### 2019-09-06 #1706 @mkearney
I've completed my review, @arokem. Following @nkratzke's lead, I also posted the following as a general remarks issue and created individual Github issues for bullet points #2-4.

Overall, I think the tool is absolutely fantastic! It's relatively easy to setup, but, more importantly, once it's up and running it's a breeze to use–e.g., in a maner of seconds, tracking the high frequency mentions of `@realDonaldTrump` generated 15+ pages of results that smoothly exported as a tidy CSV file.

With that said, I agree with the issues raised by @nkratzke and would like to add the following issues based on my review of BotSlayer:

* [x] Edit wording about tweepy functionality in README (update: [this has already been addressed](nkratzke))
* [ ] Given the real-time and interactive nature of *BotSlayer* and its relationship to public-opinion/affairs, **I think the intended audience (currently phrased as *'researchers'* or *'academic researchers'* in `paper.md` and `README.md`) should also include *'journalists'*–who not only have an interest in this topic but also have an [entire organization dedicated to computer-assisted reporting](https://www.ire.org/nicar)**.
* [ ] The tool exports `screen_name` as the user/account identifier, but screen names can be changed and are thus not stable identifiers. **The exported output should therefore include the user `id_str` value returned by Twitter's stream API**.
* [ ] The tool does not currently output a useful identifier of statuses. The exported output includes bot scores, screen names, status texts, timestamps, etc., but Twitter's stream and REST APIs [forcefully] return a lot more data than that. I don't think BotSlayer should include/return all of this information for the user, but it should at least return the sufficient information for researchers to lookup the rest via requests to the *lookup/statuses* endpoint of Twitter's REST API. **The exported output should therefore return the status `id_str` value**.
