### 2019-09-09 IUNetSci/BotSlayer-CE#16 @marcmccarty
@nkratzke How is this?:
https://github.com/IUNetSci/BotSlayer-CE/blob/master/API%20Documentation.md
