### 2019-09-09 IUNetSci/BotSlayer-CE#16 @nkratzke
Yes great, that heads into the right direction. I wonder whether the API intends to be REST? In that case, the HTTP methods should be stated. However, it seems that all API routes are GET requests? Am I right? So, it is not a REST API (which is a bit sad).

If the API is REST at least the `change_password()`, `config_ini_save()` methods would need POST/PUT/PATCH pendents.

What is more, both methods should have parameters (the properties they change). But there are no parameters described. This is strange, regardless whether the REST style is followed or not.

Furthermore, I am not quite sure why the `jsonify_user_settings()` method is included? It seems to me simply a helper method for `config_read_ep()` but is not exposed to the API itself? If that is correct it is a private helper method and should not be exposed to the API documentation (it is not callable or is it?).
