### 2019-09-09 IUNetSci/BotSlayer-CE#16 @marcmccarty
The API is not technically REST. It is not intended for independent consumption, but rather just to be used as an interface between the frontend and the middleware/backend.\
It wasn't intended to be used independently, which is why we didn't design it to conform to REST standards.

I added the parameters that are passed.

I removed the helper function.
