### 2019-09-09 IUNetSci/BotSlayer-CE#14 @yangkcatiu
Fix the postgres version to 11.5 in the dockerfile: https://github.com/IUNetSci/BotSlayer-CE/commit/90a51d059a822f830cccd133648baa27165bb2e5
