### 2019-09-09 IUNetSci/BotSlayer-CE#15 @yangkcatiu
0caaa367478caec334a82de5abfbc3ba13e48864 adds a real world use case and instructions for configure BotSlayer
