### 2019-09-09 IUNetSci/BotSlayer-CE#16 @yangkcatiu
d1d04cef53f22d0d8a00681bc8dae80610f2ce7f updates the system architecture illustration.
