### 2019-09-10 IUNetSci/BotSlayer-CE#18 @yangkcatiu
284a2d9fe45966cc7638aed48df3d4f5cf456249 community guidelines added.
