### 2019-09-10 IUNetSci/BotSlayer-CE#18 @nkratzke
Well done. Issue can be closed according to my evaluation.
