### 2019-09-10 IUNetSci/BotSlayer-CE#14 @nkratzke
Perfect. Can be closed.
