### 2019-09-10 IUNetSci/BotSlayer-CE#16 @nkratzke
Except some typos in the API documentation, I think your work improved the documentation. Thank you very much. I will close this issue. However, a proof-reading of the API documentation should be done.
