### 2019-09-10 IUNetSci/BotSlayer-CE#15 @nkratzke
Great. Really an improvement. However, I recommend adding a section regarding the Twitter access tokens (you named them APP key which is not perfectly the Twitter terminology as far as I know).

In your screenshot these fields are empty and these fields are hardly mentioned in your explaining text (which might be irritating for novice users). It is likely the most tricky part to generate and figure out these tokens for a novice user. This could simply be done, placing a link to the authoritative [Twitter Developers page](https://developer.twitter.com/en/docs/basics/authentication/guides/access-tokens.html) explaining what these tokens are, how they can be generated and used.
