### 2019-09-10 IUNetSci/BotSlayer-CE#15 @marcmccarty
We could add it to the readme, but that exact link you sent exists here, in `Help.vue` (the Help page on the frontend):
https://github.com/IUNetSci/BotSlayer-CE/blob/master/frontend/src/views/Help.vue#L17

If you hover over or click the [?] blocks in the Config page, they tell you that more info on twitter keys can be found in the help page. And then there are 3 buttons that lead to different sections of the twitter developer docs, one of which is that page you sent.
