### 2019-09-10 IUNetSci/BotSlayer-CE#15 @nkratzke
OK, you are right.
