### 2019-09-10 #1672 @jboynyc
> wave @jboynyc : would you be willing to review this article for the [Journal of Open Source Software](https://joss.theoj.org/)?

If you still need a review, I'll take a look.
