### 2019-09-17 #1706 @arokem
@mkearney, @nkratzke : could you please take a look when you get a chance? Were your previous comments addressed through these revisions?
