### 2019-09-18 #1706 @arokem
@nkratzke : could you please check the boxes at the top of this issue?
