### 2019-09-24 #1706 @arokem
@mkearney : I see that there are still some check-boxes that you did not check. Do the authors need to address any of these before you can check these off? Or were these issues addressed in the previous round of reviews?
