### 2019-10-01 #1706 @arokem
@mkearney : have you had a chance to take a look and see whether you are able to check off the remaining items?
