### 2019-10-15 #1706 @arokem
I apologize for how long this has languished. I've contacted @mkearney by email and he told me that he has some time set aside for this tonight, so I hope we can put this thing to rest tomorrow...
