### 2019-10-15 #1706 @mkearney
Sorry for the delay–you all were so fast with the revisions! The revisions look great!
