### 2019-10-15 #1706 @arokem
OK - I see that you have checked all the boxes. I take it to assume that both reviewers are ready to recommend acceptance at this point. Thank you both for your reviews!
