### 2019-10-15 #1706 @arokem
@yangkcatiu : see ☝️above for missing DOIs in references. Could you please add these for these references? I will also take another read through the PDF to make sure the everything is in the right place.
