### 2019-10-15 #1706 @arokem
I noticed that there is some inconsistent reference formatting. For example, sometimes "Clayton Allen Davis et al., 2016", sometimes "C A Davis et al. 2016" , and sometimes  "Davis et al. 2016". Could you please make sure that reference formats are consistent?
