### 2019-10-15 #1706 @yangkcatiu
I have added the missing references; I also updated the references for two recently updated papers.
