### 2019-10-15 #1706 @arokem
Looks like you got all the DOIs. Did you see my formatting PR? https://github.com/IUNetSci/BotSlayer-CE/pull/37
