### 2019-10-15 #1706 @yangkcatiu
I have re-formated the references.
