### 2019-10-16 #1706 @arokem
I had one more suggested wording change in https://github.com/IUNetSci/BotSlayer-CE/pull/38. If that looks OK, feel free to merge that.

At this point, could you also please make a new release of the main repo that includes all of the changes that have resulted from the review? Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the archive version, please make sure that:

- The authors of the archive are the same as the JOSS paper authors

Once we've registered the archive DOI, I believe that the paper will be ready for final approval by the EIC.
