### 2019-10-17 #1706 @Kevin-Mattheus-Moerman
@yangkcatiu Below are some comments and suggestions on the paper:

- [ ] Instead of `First` and `Second`, I suggest `Firstly` and `Secondly`.
- [ ] Is it possible to increase the font in the table view of figure 1?
- [ ] Add `s` to make calculations plural in "All statistical and machine learning
calculation[s]..."
- [ ] Please define the acronym AUC
