### 2019-10-17 #1706 @yangkcatiu
@Kevin-Mattheus-Moerman  thank you for your comments! I have made the changes accordingly:

- [x] `First` -> `Firstly`; `Second` -> `Secondly`
- [x] Increase the font size in the table view of figure 1
- [x] Add `s` to `calculations`
- [x] Define AUC
