### 2019-10-18 #1706 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
